requirements.txt
django_badges==1.0.0
Django==1.11.4
django_ace==1.0.3
